+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.3.0](https://github.com/studio24/amplify/compare/v1.2.1...v1.3.0) (2023-08-08)
+
+
+### Features
+
+* add sortable tables ([#28](https://github.com/studio24/amplify/issues/28)) ([1a22590](https://github.com/studio24/amplify/commit/1a22590f6adfcd880eb043cf544d9ea641305de0))
+
+## [1.2.1](https://github.com/studio24/amplify/compare/v1.2.0...v1.2.1) (2023-03-30)
+
+
+### Bug Fixes
+
+* update NPM packages ([a27f8c1](https://github.com/studio24/amplify/commit/a27f8c1819a405a2551c13c1bba462b4b6148873))
+
+## [1.2.0](https://github.com/studio24/amplify/compare/v1.1.0...v1.2.0) (2023-03-30)
+
+
+### Features
+
+* add navigation patterns ([#14](https://github.com/studio24/amplify/issues/14)) ([caeb922](https://github.com/studio24/amplify/commit/caeb922277c29b82c036f058a74c67e074c22606))
+
+## [1.1.0](https://github.com/studio24/amplify/compare/v1.0.1...v1.1.0) (2022-12-21)
+
+
+### Features
+
+* carousel + mixin fix ([6d369d4](https://github.com/studio24/amplify/commit/6d369d4031b694cddde061021f6aeaafac9729f7))
+
 ## [1.0.1](https://github.com/studio24/amplify/compare/v1.0.0...v1.0.1) (2022-12-20)